requirements.txt
numpy
pandas
scikit-learn
xgboost
fastapi
uvicorn
pydantic
joblib
pyyaml
